requirements.txt
pandas 
numpy
seaborn
matplotlib
ipykernel
mysql-connector-python
sqlalchemy
ipython-sql         